Comman hash step base class extracted.
